adds a generator command which discover providers
This command will be triggered by the go:generate annotation,
registry that need to auto discover their components need to be
defined as follow:
```
//go:generate go run github.com/BetterWorks/webhook-engine/cmd/generator
type Registry struct {
	di.EmptyRegistry
}
```
This will trigger the generator, this generator will find providers
and the location of the registry to use (a `Register` method
implementation will be generated containing all the registrations).

For now we are only gathering information, now we just need to output
those information into a generated file.
In the future we might want to parse more information, like the name
of the dependencies, ...