allows injection multiple
Because this is an awesome feature, to get all routes for a router,
or all X for a manager of Xs...

And in order to do that a consequent refactor has been made to the
resolver, with a single resolve method, as slice, map, instance of T
in the reflection world are all `reflect.Value`, so one method to
rule them all.

Then we have a separation of concern, instead of resolving queries,
we resolve request. And a request is a combination of two distinct
things:
- a query => responsibility: finding provider definition
- a collector => responsibility: checking if the found providers are
  corresponding to the collector constraints (unique, multiple,
mandatory, ...) and also building the `reflect.Value`.

All of that seems way more cleaner and shape things to implement
optional dependency quite quickly for example.