introduces dynamic providers
This is a way to provides providers whenever we look for one,
because up front we might sometimes be able to build a lot, and we
don't need to register them if no one is requesting for them.

This is for example the case of env vars, now we can register a
dynamic provider which will provides the env var, but only if one is
requesting them.

This is a bit a way to have a provider saying, I can produce "foo"
or "bar" or "...".

We should be able to re-visit what was done with the config, and not
need the AST no more (which will help to boost the generation), one
dynamic provider might be suitable to achieve the same thing we've
did.