adds a config dynamic provider
Allowing to provides all the fields from a config struct. Thanks to
this we will be able to remove all the providers that we were
computing during the AST of the config type analysis. It will clean
up the "auto register" method.

One thing, as dynamic provider are returning a provider function
which will be parsed during the `Register` method of the resolver,
we need to have the perfect return type. But this dynamic provider
is a good example of a provider being able to return various types.
So we managed to solve this using reflection and building the
correct signature. But this is not convenient, maybe we should have
a provider type which does use "discovery", but is explicit about
what it is returning.
Like that we could return that provider which would have a factory
method returning any.