adds capability to describe the resolver
Which might be handy to debug as this is listing all providers and
every components available in the store.

Also we were capturing the description while parsing factory methods
annotated with `@provider` but we weren't doing anything with it.
Now when registering a factory method we can do
`godi.Description(...)`. This is what is automatically generated by
the parser.