makes register/resolution be thread safe
For register:
- use a "copy on write" sorted structure, we have way more read than
  write, so that probably make sense to use a "lock free" structure
even if it involves a small penalty on write.
For resolution:
- use a lock for the name to be build, when building a name, first
  generate a lock for the name and use it. Like that if we have two
concurrent build, the first will generate the lock for the name and
lock it, then the second one will find the existing lock and wait
for the lock to be released (once the component is built). Then it
will search if the component is available in the store, and use it
if it is.
We won't keep the lock in the map (name/lock) after the build, as
this is unlikely that someone will try to build again the same
component as it is available in the store.

Note that we will probably be lock free for transient provider when
we will introduce the scope concept. So we will keep the lock only
for singleton scope (most of the providers)