adds initializers
Because sometimes we want to do something before starting the app,
like initializing some global level on the logger, or else...
One can provide providers providing empty function (returning error
or not, so unsafe initializer or initializer).

When we will run `Initialize` on the resolver, all those
"initializers" will be executed.